Faq Entry about dmenu 9menu etc
Hi all,

it has been said on the #awesome, a Faq entry about dmenu etc
would be nice.

I just added something there, in the way it have been asked.

I don't have ikiwiki actually and was not able to check if it
results in a proper html page.

Also I didn't think about if it makes sense or not, so may ignore
it..