page template: Use BASEURL variable
Instead of using absolute URLs, this now uses an ikiwiki variable to get
relative paths. For example, in index.html this will expand to the empty
string while in download/index.html it expands to "../".

Since this expands to the empty string on index.html, some things which
previously linked to / will now have to link to index.html. However,
that shouldn't be a problem, I think.

Signed-off-by: Uli Schlachter <psychon@znc.in>