Change the appearance of the top-menu
When you hover over one of the items in the top menu, e.g. "home", a
background appears. This commit changes this background to cover the
full height of the top-menu instead of only a small rectangle around the
item.

The "margin: 0px" is needed so that this background can "reach" up to
the top of the page. Note that this however also affects other elements
on the page and now the "awesome" at the top left of the page is a bit
farther up.

The height of "#top" and the padding around the links then change the
actual height of this background.

This is partial revert of 34c806d8176a87398e2372.

Signed-off-by: Uli Schlachter <psychon@znc.in>