Add a doing_a_release page
This page is described on the old document that was available on the
wiki. However, I updated it to the new state of things and, for example,
left out the mailing list. This means that lots of the things here are
new and untested, but I guess they will be tested soon.

Signed-off-by: Uli Schlachter <psychon@znc.in>