Fix "fix xrandr.lua link"
It is my understanding that [title](target) is meant for external links
while [[title|target]] does internal links. The main difference here is
that ikiwiki automatically finds the target and makes it work with its
directory structure, so that no "manual ../" is needed.

This amends/fixes commit f9e7311e.

Signed-off-by: Uli Schlachter <psychon@znc.in>